migrate the environment back to production
