+++ Bot/Moderation/Message_Checks/__init__.py
@@ -1,15 +1,9 @@
 
-#from Moderation.Message_Checks import banned_word, emoji_spam, illegal_char, invite, link, mention_spam, spam
-from Moderation.Message_Checks import emoji_spam, illegal_char, invite, link, mention_spam, spam
+from Moderation.Message_Checks import banned_word, emoji_spam, illegal_char, invite, link, mention_spam, spam
 
-# checks = [
-#    banned_word, spam,
-#    emoji_spam, illegal_char,
-#    invite, link, mention_spam
-# ]
 
 checks = [
-    spam,
+    banned_word, spam,
     emoji_spam, illegal_char,
     invite, link, mention_spam
 ]
